move storage and cluster operations to saga (#721)
* move storage and cluster operations to saga

* lint

* remove thunk deps
