+++ src/app/cluster/duck/index.ts
@@ -1,5 +1,5 @@
 import clusterReducer from './reducers';
-export { default as clusterOperations } from './operations';
+export { default as clusterSagas } from './sagas';
 export { ClusterActionTypes as clusterTypes } from './actions';
 export { default as clusterSelectors } from './selectors';
 export default clusterReducer;
