+++ src/app/home/DetailViewComponent.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from 'react';
 import { connect } from 'react-redux';
-import clusterOperations from '../cluster/duck/operations';
-import storageOperations from '../storage/duck/operations';
+import clusterSagas from '../cluster/duck/sagas';
+import storageSagas from '../storage/duck/sagas';
 import clusterSelectors from '../cluster/duck/selectors';
 import storageSelectors from '../storage/duck/selectors';
 import planSelectors from '../plan/duck/selectors';
@@ -191,10 +191,10 @@ function mapStateToProps(state) {
 
 const mapDispatchToProps = dispatch => {
   return {
-    removeCluster: id => dispatch(clusterOperations.removeCluster(id)),
-    removeStorage: id => dispatch(storageOperations.removeStorage(id)),
     runStageRequest: plan => dispatch(PlanActions.runStageRequest(plan)),
     runMigrationRequest: (plan, disableQuiesce) => dispatch(PlanActions.runMigrationRequest(plan, disableQuiesce)),
+    removeCluster: name => dispatch(ClusterActions.removeClusterRequest(name)),
+    removeStorage: name => dispatch(StorageActions.removeStorageRequest(name)),
     updateStageProgress: (plan, progress) =>
       dispatch(PlanActions.updateStageProgress(plan.planName, progress)),
     stagingSuccess: plan => dispatch(PlanActions.stagingSuccess(plan.planName)),
