+++ src/app/home/duck/index.ts
@@ -1,6 +1,2 @@
-// import homeHooks from './hooks';
-// export { default as clusterOperations } from './operations';
-// export { Types as clusterTypes } from './actions';
 export { default as homeHooks } from './hooks';
 export { default as homeContext } from './context';
-// export default clusterReducer;
