+++ src/app/storage/duck/index.ts
@@ -1,5 +1,5 @@
 import storageReducer from './reducers';
-export { default as storageOperations } from './operations';
+export { default as storageSagas } from './sagas';
 export { StorageActionTypes as storageTypes } from './actions';
 export { default as clusterSelectors } from './selectors';
 export default storageReducer;
