+++ src/sagas.ts
@@ -54,9 +54,11 @@ export default function* rootSaga() {
     logSagas.watchLogsPolling(),
     logSagas.watchLogDownload(),
     logSagas.watchLogsDownload(),
+    clusterSagas.watchRemoveClusterRequest(),
     clusterSagas.watchAddClusterRequest(),
     clusterSagas.watchUpdateClusterRequest(),
     clusterSagas.watchClusterAddEditStatus(),
+    storageSagas.watchRemoveStorageRequest(),
     storageSagas.watchAddStorageRequest(),
     storageSagas.watchStorageAddEditStatus(),
     storageSagas.watchUpdateStorageRequest(),
